% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_to_grid}
\alias{df_to_grid}
\title{Replace raster values by a column of a data frame}
\usage{
df_to_grid(data, grid, column, ..., spat_form = ~longitude + latitude,
  reset_outside = TRUE, outside_value = NA)
}
\arguments{
\item{data}{The data set that should be embeded by grid data. The
grid data is specified via \code{file = } argument. See \code{read_grid}
for details.}

\item{grid}{A Raster object or file to a GeoTIFF file that can be read in
via \code{raster}.}

\item{...}{Further arguments passed to \code{read_grid} if \code{grid} is
a file instead of a raster object.}

\item{spat_form}{A \code{formula} object. Specifies the longitude and
latitude column names in \code{data}. These will be used in combination with
\code{link[raster]{extract}} to embed \code{data} with grid data efficiently.}
}
\description{
Replaces values of \code{grid} with specified column at the coordinates of
raster that correspond to coordinates in \code{data} (column containing
coordinates are specified via \code{spat_form}). If \code{outside_na = TRUE}
(the default), \code{grid} values at coordinates outside of coordinates
in \code{data} are set to \code{NA}.
}
\keyword{internal}
