% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{get_coords}
\alias{get_coords}
\title{Extract coordinates from data and turn into SpatialPointsDataFrame}
\usage{
get_coords(data, spat_form, proj)
}
\description{
Extract coordinates from data and turn into SpatialPointsDataFrame
}
\keyword{internal}
