% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_extent}
\alias{get_extent}
\title{Define extent}
\usage{
get_extent(extent = NULL, mask = NULL)
}
\description{
Depending on input, creates a \code{\link[raster]{extent}} object,
either from specified extent, the extent of a mask, or sets extent to the
maximum extent of a geographic map.
}
\keyword{internal}
