% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-newdata.R
\name{grid_to_df}
\alias{grid_to_df}
\alias{grid_as_coord_df}
\title{Creates a data frame from grid (file)}
\usage{
grid_to_df(grid, extent = NULL, mask = NULL, spat_form = ~longitude +
  latitude, na.rm = TRUE)

grid_as_coord_df(grid, extent = NULL, mask = NULL,
  spat_form = ~longitude + latitude)
}
\arguments{
\item{spat_form}{A \code{formula} object. Specifies the longitude and
latitude column names in \code{data}. These will be used in combination with
\code{link[raster]{extract}} to embed \code{data} with grid data efficiently.}
}
\description{
Creates a data frame from grid file, having as many rows as unique coordinates
in the grid file, excluding cells with \code{NA} values.
}
