% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{group_files_by_year}
\alias{group_files_by_year}
\title{Group files by year}
\usage{
group_files_by_year(queries, years, ...)
}
\description{
See \code{match_file_by year} for details.
}
\examples{
queries <- list(
 search_files("acces.*weiss"),
 search_files("class00", "adj"))
mastergrids:::group_files_by_year(queries, years = 1997:2003)
}
\keyword{internal}
