% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searching.R
\name{list_paths}
\alias{list_paths}
\title{List mastergrid paths}
\usage{
list_paths()
}
\description{
A data frame containing names of mastergrid paths available in the
\code{mastergrid_paths} object and the respective values.
}
