% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mastergrid_paths}
\alias{mastergrid_paths}
\title{Some predifined folder paths in the MAP mastergrids folder}
\format{An object of class \code{list} of length 16.}
\usage{
mastergrid_paths
}
\description{
A collection of mastergrid paths that can be used to narrow the search path
of \code{search_files}.
}
\keyword{datasets}
