% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{match_file_to_year}
\alias{match_file_to_year}
\title{Find the file that matches the year most closely}
\usage{
match_file_to_year(files, year)
}
\description{
Find the file that matches the year most closely
}
\keyword{internal}
