% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_names}
\alias{process_names}
\title{Process names}
\usage{
process_names(names, files, data, grids)
}
\description{
If provided, performs sanity checks, otherwise extracts names from
names of the raster layers provided in the \code{grid} argument.
}
\keyword{internal}
