% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-grid.R
\name{read_grids}
\alias{read_grids}
\title{Import multiple grid data sets}
\usage{
read_grids(files, extent = NULL, mask = NULL, ...,
  .progress = interactive())
}
\arguments{
\item{files}{Full file paths to grid data.}

\item{extent}{An object of class \code{\link[raster]{extent}}.
The spatial # extent of interest. Only the part of the grid
data within \code{extent} will be read in. If \code{NULL}, extent will be
determined from \code{mask}. If both NULL, the whole grid file will be read
in. Used as second argument in \code{\link[raster]{crop}}.}

\item{mask}{An object used to define a mask. All values of the
(cropped) grid outside the \code{mask} will be set to \code{NA}.
See \code{\link[raster]{mask}} for details.}

\item{...}{Currently not used}

\item{.progress}{A logical, for whether or not to print a progress bar for
multiprocess, multisession, and multicore plans.}
}
\description{
Import multiple grid data sets
}
