% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searching.R
\name{search_files}
\alias{search_files}
\title{List all files in directory}
\usage{
search_files(..., path = "/media/z/mastergrids/", full.names = TRUE,
  recursive = TRUE, file_extension = ".tif$")
}
\arguments{
\item{...}{Comma separated character strings that define parts of the file
parts defined in \code{...}. Specifications can also contain regular expressions.}

\item{path}{a character vector of full path names; the default
    corresponds to the working directory, \code{\link{getwd}()}.  Tilde
    expansion (see \code{\link{path.expand}}) is performed.  Missing
    values will be ignored.}

\item{full.names}{a logical value.  If \code{TRUE}, the directory
    path is prepended to the file names to give a relative file path.
    If \code{FALSE}, the file names (rather than paths) are returned.}

\item{recursive}{logical.  Should the listing recurse into directories?}

\item{file_extension}{Only files with this extension will be returned.
Usually we are only interested in GeoTiff files, thus \code{".tif$"} is
set by default.}
}
\description{
List all files in directory
}
\examples{
library(mastergrids)
# get mastergrids accessibility data (friction 2015)
search_files("friction", path = mastergrid_paths$accessibility_weiss)
search_files("Landcover", "Class02", "2005", "percentage",
  path = mastergrid_paths$landcover)
# search the entire mastergrid folder:
system.time({
  search_files("LST_Night", "2002", "Annual.max.5km")
})
}
