% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_extract_year}
\alias{str_extract_year}
\title{Extract year from character (vector)}
\usage{
str_extract_year(x)
}
\arguments{
\item{x}{A character (vector).}
}
\description{
Given file path(s), extracts the year via regular expression and coverts
to integer.
}
\keyword{internal}
